backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
azure-ai-agents
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.9.0
azure-identity==1.23.0
azure-search-documents==11.5.3
azure-storage-blob==12.25.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.5
marshmallow==3.26.1
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.0.2
openai==1.96.0
openpyxl==3.1.5
orjson==3.11.0
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
pip==23.0.1
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
PyJWT==2.10.1
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
pdfkit
beautifulsoup4
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==58.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
locust
fastapi
uvicorn
openai
azure-search-documents
azure-core
python-dotenv
pymupdf
python-docx

